+++ app/components/play/play-log-display.tsx
@@ -34,7 +34,7 @@ export function PlayLogDisplay({
 
   return (
     <ScrollArea
-      className="min-h-0 w-full"
+      className="play-log-texture min-h-0 w-full"
       viewportRef={viewportRef}
       onViewportScroll={onViewportScroll}
       viewportClassName="!flex !flex-col"